demo_target/requirements.txt
flask==3.0.0
groq==1.1.2
chromadb==1.5.5
python-dotenv==1.0.0
markupsafe==2.1.3